Extract audio processing display model
